+++ src/mol-gl/shader/chunks/apply-interior-color.glsl.ts
@@ -1,5 +1,18 @@
 export const apply_interior_color = `
 if (interior) {
+    #if defined(dInteriorColorType_uniform)
+        #ifdef dUsePalette
+            material.rgb = texture2D(tPalette, vec2(decodePaletteV(uInteriorThemeColor), 0.5)).rgb;
+        #else
+            material.rgb = uInteriorThemeColor;
+        #endif
+    #elif defined(dInteriorColorType_texture)
+        #ifdef dUsePalette
+            material.rgb = texture2D(tPalette, vec2(vInteriorPaletteV, 0.5)).rgb;
+        #else
+            material.rgb = vInteriorColor;
+        #endif
+    #endif
     material.rgb = mix(material.rgb, uInteriorColor.rgb, uInteriorColor.a);
 
     float isf = clamp(uInteriorSubstance.a, 0.0, 0.99); // clamp to avoid artifacts