+++ src/mol-repr/particles/visual.ts
@@ -20,6 +20,8 @@ import { ColorTheme } from '../../mol-theme/color';
 import { ValueCell } from '../../mol-util';
 import { createSizes } from '../../mol-geo/geometry/size-data';
 import { createColors } from '../../mol-geo/geometry/color-data';
+import { updateInteriorColors } from '../../mol-geo/geometry/interior-color-data';
+import { hasInteriorThemeColor } from '../../mol-geo/geometry/interior';
 import { MarkerAction } from '../../mol-util/marker-action';
 import { Mat4 } from '../../mol-math/linear-algebra';
 import { Overpaint } from '../../mol-theme/overpaint';
@@ -237,6 +239,7 @@ export function ParticleVisual<G extends Geometry, P extends ParticleParams & Ge
 
             if (updateState.updateColor) {
                 createColors(locationIt, positionIt, effectiveTheme.color, renderObject.values);
+                updateInteriorColors(renderObject.values, locationIt, positionIt, effectiveTheme.color, hasInteriorThemeColor(newProps));
             }
 
             updateValues(renderObject.values, newProps);