+++ src/mol-repr/structure/complex-visual.ts
@@ -21,6 +21,8 @@ import { ColorTheme } from '../../mol-theme/color';
 import { ValueCell, deepEqual } from '../../mol-util';
 import { createSizes, SizeData } from '../../mol-geo/geometry/size-data';
 import { createColors } from '../../mol-geo/geometry/color-data';
+import { updateInteriorColors } from '../../mol-geo/geometry/interior-color-data';
+import { hasInteriorThemeColor } from '../../mol-geo/geometry/interior';
 import { MarkerAction } from '../../mol-util/marker-action';
 import { Mat4 } from '../../mol-math/linear-algebra';
 import { Overpaint } from '../../mol-theme/overpaint';
@@ -124,6 +126,10 @@ export function ComplexVisual<G extends Geometry, P extends StructureParams & Ge
             updateState.updateColor = true;
         }
 
+        if (hasInteriorThemeColor(newProps) !== hasInteriorThemeColor(currentProps)) {
+            updateState.updateColor = true;
+        }
+
         if (!SizeTheme.areEqual(theme.size, currentTheme.size)) {
             updateState.updateSize = true;
         }
@@ -205,6 +211,7 @@ export function ComplexVisual<G extends Geometry, P extends StructureParams & Ge
 
             if (updateState.updateColor) {
                 createColors(locationIt, positionIt, newTheme.color, renderObject.values);
+                updateInteriorColors(renderObject.values, locationIt, positionIt, newTheme.color, hasInteriorThemeColor(newProps));
             }
 
             updateValues(renderObject.values, newProps);